refactor : interceptorURL 정리
