+++ src/main/java/space/space_spring/dto/space/GetSpaceHomeDto.java
@@ -7,9 +7,6 @@ import space.space_spring.dto.pay.dto.PayReceiveInfoDto;
 import space.space_spring.dto.pay.dto.PayRequestInfoDto;
 import java.util.List;
 
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
 public class GetSpaceHomeDto {
 
     @Getter
